feat(accessIpPie): adds location data for selected ip
